+++ lib/data/tzdb/latest_10y.dart
@@ -5,7 +5,7 @@ import 'package:time_machine2/src/platforms/dart_native_io.dart';
 
 void registerTzdbData_latest_10y() {
   TimeMachineIO.registerAsset('tzdb', 'latest_10y.tzf',
-      ByteData.sublistView(Uint8List.fromList(_embeddedData.codeUnits)));
+      ByteData.sublistView(Uint16List.fromList(_embeddedData.codeUnits)));
 }
 
 const _embeddedData =
